+++ src/types.ts
@@ -40,6 +40,7 @@ export enum WeightName {
     NUM_RETOOTS = 'NumRetoots',
     RETOOTED_IN_FEED = 'RetootedInFeed',
     TIME_DECAY = 'TimeDecay',
+    TRENDING_LINKS = 'TrendingLinks',
     TRENDING_TAGS = "TrendingTags",
     TRENDING_TOOTS = "TrendingToots",
     VIDEO_ATTACHMENTS = 'VideoAttachments',
@@ -128,6 +129,7 @@ export type UserData = {
     followedTags: StringNumberDict,
     mutedAccounts: AccountNames,
     serverSideFilters: mastodon.v2.Filter[],
+    trendingLinks: mastodon.v1.TrendLink[],
 };
 
 export interface TootExtension extends mastodon.v1.Status {
@@ -148,6 +150,14 @@ export type TootScore = {
     weightedScores: Weights;
 };
 
+export interface TrendingLink extends mastodon.v1.TrendLink {
+    numToots?: number;
+    numAccounts?: number;
+};
+
+export type TrendingLinkUrls = Record<string, TrendingLink>;
+// export type TrendingLinkUrls = Record<mastodon.v1.TrendLink["url"], TrendingLink>;
+
 export interface TrendingTag extends mastodon.v1.Tag {
     numAccounts?: number;
     numToots?: number;
@@ -156,7 +166,7 @@ export interface TrendingTag extends mastodon.v1.Tag {
 
 export type StorageValue = FeedFilterSettings | FeedFilterSettingsSerialized | StringNumberDict |
     TootExtension[] | TootURIs | Weights | mastodon.v1.Account | mastodon.v1.Account[] | mastodon.v2.Filter[] |
-    number;
+    mastodon.v1.TrendLink[] | number;
 
 
 // From https://dev.to/nikosanif/create-promises-with-timeout-error-in-typescript-fmm